% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magreport-package.R
\docType{package}
\name{magreport-package}
\alias{magreport}
\alias{magreport-package}
\title{magreport: Explore and Report MAG Recovery Results}
\description{
Reads the output tables of the mag-pipeline metagenomics workflow (assembly statistics, read mapping, bin statistics, CheckM2 quality and GTDB-Tk taxonomy) into tidy data frames, and summarises MAG recovery following the MIMAG quality standards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AntonioPuriel/magreport}
  \item Report bugs at \url{https://github.com/AntonioPuriel/magreport/issues}
}

}
\author{
\strong{Maintainer}: Antonio Puriel Hernández \email{apurielh@gmail.com} (\href{https://orcid.org/0009-0000-7631-1102}{ORCID})

Authors:
\itemize{
  \item Antonio Puriel Hernández \email{apurielh@gmail.com} (\href{https://orcid.org/0009-0000-7631-1102}{ORCID})
}

}
\keyword{internal}
