% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quality_colours}
\alias{quality_colours}
\title{Colours used for the quality categories}
\usage{
quality_colours()
}
\value{
A named character vector of colours, one per level of
\code{\link[=quality_levels]{quality_levels()}}.
}
\description{
Colours used for the quality categories
}
\examples{
quality_colours()
}
